IGNITE-23126 Move type handling to separate file
