+++ modules/platforms/python/cpp_module/py_cursor.cpp
@@ -18,96 +18,16 @@
 #include <ignite/odbc/sql_statement.h>
 #include <ignite/odbc/query/data_query.h>
 
-#include <ignite/common/detail/config.h>
-
 #include "module.h"
 #include "py_cursor.h"
 #include "py_sequence.h"
+#include "type_conversion.h"
 
 #include <Python.h>
 
-static PyObject* primitive_to_pyobject(ignite::primitive value) {
-    using ignite::ignite_type;
-
-    if (value.is_null()) {
-        Py_INCREF(Py_None);
-        return Py_None;
-    }
-
-    switch (value.get_type()) {
-        case ignite_type::STRING: {
-            auto &str_val = value.get<std::string>();
-            return PyUnicode_FromStringAndSize(str_val.c_str(), str_val.size());
-        }
-
-        case ignite_type::INT8: {
-            auto &i8_val =  value.get<std::int8_t>();
-            return PyLong_FromLong(long(i8_val));
-        }
-
-        case ignite_type::INT16: {
-            auto &i16_val =  value.get<std::int16_t>();
-            return PyLong_FromLong(long(i16_val));
-        }
-
-        case ignite_type::INT32: {
-            auto &i32_val =  value.get<std::int32_t>();
-            return PyLong_FromLong(long(i32_val));
-        }
-
-        case ignite_type::INT64: {
-            auto &i64_val =  value.get<std::int64_t>();
-            return PyLong_FromLongLong(i64_val);
-        }
-
-        case ignite_type::FLOAT: {
-            auto &float_val =  value.get<float>();
-            return PyFloat_FromDouble(float_val);
-        }
-
-        case ignite_type::DOUBLE: {
-            auto &double_val =  value.get<double>();
-            return PyFloat_FromDouble(double_val);
-        }
-
-        case ignite_type::BOOLEAN: {
-            auto &bool_val =  value.get<bool>();
-            if (bool_val) {
-                Py_RETURN_TRUE;
-            } else {
-                Py_RETURN_FALSE;
-            }
-        }
-
-        case ignite_type::BYTE_ARRAY: {
-            auto &blob_val =  value.get<std::vector<std::byte>>();
-            return PyBytes_FromStringAndSize((const char*)blob_val.data(), blob_val.size());
-        }
-
-        case ignite_type::UUID:
-        case ignite_type::DATE:
-        case ignite_type::TIMESTAMP:
-        case ignite_type::TIME:
-        case ignite_type::DATETIME:
-        case ignite_type::BITMASK:
-        case ignite_type::DECIMAL:
-        case ignite_type::PERIOD:
-        case ignite_type::DURATION:
-        case ignite_type::NUMBER:
-        default: {
-            // TODO: IGNITE-22745 Provide wider data types support
-            auto err_msg = "The type is not supported yet: " + std::to_string(int(value.get_type()));
-            PyErr_SetString(PyExc_RuntimeError, err_msg.c_str());
-            return nullptr;
-        }
-    }
-}
-
-static void write_pyobject(ignite::protocol::writer &writer, PyObject *obj) {
-
-}
-
-
+/**
+ * Python parameter set.
+ */
 class py_parameter_set : public ignite::parameter_set {
 public:
     /**
@@ -393,7 +313,7 @@ static PyObject* py_cursor_column_type_code(py_cursor* self, PyObject* args)
     return PyLong_FromLong(long(column->get_data_type()));
 }
 
-static PyObject* py_cursor_column_display_size(py_cursor* self, PyObject* args)
+static PyObject* py_cursor_column_display_size(py_cursor* self, PyObject*)
 {
     if (!self->m_statement) {
         PyErr_SetString(PyExc_RuntimeError, "Cursor is in invalid state (Already closed?)");
@@ -404,7 +324,7 @@ static PyObject* py_cursor_column_display_size(py_cursor* self, PyObject* args)
     return Py_None;
 }
 
-static PyObject* py_cursor_column_internal_size(py_cursor* self, PyObject* args)
+static PyObject* py_cursor_column_internal_size(py_cursor* self, PyObject*)
 {
     if (!self->m_statement) {
         PyErr_SetString(PyExc_RuntimeError, "Cursor is in invalid state (Already closed?)");
